Show nothing if nothing is to be shown
Retreats from an earlier commit where it would show blank.
To prevent the case of misattribution it is better to not attempt
to show a name if nothing is to be quoted.

Also, reduces inducing that "huh?" moment by returning right back
into the command prompt instead of a jump over a block of newlines
for the sake of nothing.