Instruct to copy instead of moving script
Copying out the script in the first instance of trying out the
post-commit hook will at least allow for redoing the same in
another repo, or setting it as a global hook. Moving it from the
first step would lose track of where the script went in terms
of the README narrative.